+++ seleniumtid/utils.py
@@ -14,6 +14,9 @@ import os
 from seleniumtid import selenium_driver
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+import requests
+from urlparse import urlparse
+import time
 
 
 class Utils(object):
@@ -66,3 +69,91 @@ class Utils(object):
         WebDriverWait(self.driver, timeout).until(EC.invisibility_of_element_located(locator))
         # Restore implicit wait from properties
         self.set_implicit_wait()
+
+    def get_remote_node(self):
+        '''
+        Return the remote node that it's executing the actual test session
+        '''
+        logging.getLogger("requests").setLevel(logging.WARNING)
+        remote_node = None
+        if selenium_driver.config.getboolean_optional('Server', 'enabled'):
+            # Request session info from grid hub
+            host = selenium_driver.config.get('Server', 'host')
+            port = selenium_driver.config.get('Server', 'port')
+            session_id = self.driver.session_id
+            url = 'http://{}:{}/grid/api/testsession?session={}'.format(host, port, session_id)
+            response = requests.get(url).json()
+
+            # Extract remote node from response
+            remote_node = urlparse(response['proxyId']).hostname
+            self.logger.debug("The test is executing in the remote node {}".format(remote_node))
+        return remote_node
+
+    def get_remote_video_node(self):
+        '''
+        Return the remote node only if it has video capabilities
+        '''
+        remote_node = self.get_remote_node()
+        if remote_node and self._is_remote_video_enabled(remote_node):
+            # Wait to the video recorder start
+            time.sleep(1)
+            return remote_node
+        return None
+
+    def download_remote_video(self, remote_node, session_id, video_name):
+        '''
+        Download the video recorded in the remote node during the specified test session and save it in videos folder
+        '''
+        try:
+            video_url = self._get_remote_video_url(remote_node, session_id)
+        except requests.exceptions.ConnectionError:
+            self.logger.warn("Remote server seems not to have video capabilities")
+            return
+
+        if not video_url:
+            self.logger.warn("Test video not found in node '{}'".format(remote_node))
+            return
+
+        self._download_video(video_url, video_name)
+
+    def _get_remote_node_url(self, remote_node):
+        logging.getLogger("requests").setLevel(logging.WARNING)
+        gridextras_port = 3000
+        return 'http://{}:{}'.format(remote_node, gridextras_port)
+
+    def _get_remote_video_url(self, remote_node, session_id):
+        url = '{}/video'.format(self._get_remote_node_url(remote_node))
+        timeout = time.time() + 5  # 5 seconds from now
+
+        # Requests videos list until timeout or the video url is found
+        video_url = None
+        while time.time() < timeout:
+            response = requests.get(url).json()
+            try:
+                video_url = response['available_videos'][session_id]['download_url']
+                break
+            except KeyError:
+                time.sleep(1)
+        return video_url
+
+    def _download_video(self, video_url, video_name):
+        filename = '{0:0=2d}_{1}.mp4'.format(selenium_driver.videos_number, video_name)
+        filepath = os.path.join(selenium_driver.videos_path, filename)
+        if not os.path.exists(selenium_driver.videos_path):
+            os.makedirs(selenium_driver.videos_path)
+        response = requests.get(video_url)
+        open(filepath, 'wb').write(response.content)
+        self.logger.info("Test video saved in '{}'".format(filepath))
+        selenium_driver.videos_number += 1
+
+    def _is_remote_video_enabled(self, remote_node):
+        '''
+        Check if the remote node has the video recorder enabled
+        '''
+        url = '{}/config'.format(self._get_remote_node_url(remote_node))
+        try:
+            response = requests.get(url).json()
+            record_videos = response['config_runtime']['theConfigMap']['video_recording_options']['record_test_videos']
+        except (requests.exceptions.ConnectionError, KeyError):
+            record_videos = 'false'
+        return True if record_videos == 'true' else False