+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.5
 
 *Release date: 2014-11-??*
 
+  * Downloads the saved video if the test has been executed in a VideoGrid
   * Add BasicTestCase class to be used in Api tests or in other tests without selenium driver
 
 v0.4